requirements.txt
beaupy==3.5.3
commonmark==0.9.1
emoji==2.2.0
pycryptodome==3.17
Pygments==2.15.1
pystyle==2.9
python-yakh==0.3.1
rich==12.6.0
